chore: move LoggerInterface to common
